Base Functions Module
This module holds all of the SQL queries and base functions that are inherited by many other objects throughout the program. It is highly advised to jump into this file and alter the SQLQuery function Get_Database_Connect. Fill in your desired Server Name, UserName,  and Password to the MSSQl software on your host machine.